==> internal-projects/23_get-all-the-examples.R <==
## produced googlesheets-Ex.R
## a script of all examples, uncommented, ready to run
## make sure most recent version has been installed in the normal library
tools:::.createExdotR(pkg = "googlesheets",
                      file.path(.libPaths()[1], "googlesheets"),
                      commentDontrun = FALSE)
